Clean up loading and clearing session
